move to a folder centered file structure
