+++ airbyte-ci/connectors_ci/metadata_service/orchestrator/config.py
@@ -1,0 +1,4 @@
+# move to config -> metadata service
+BUCKET_NAME = "ben-ab-test-bucket"
+CATALOG_FOLDER = "catalogs"
+REPORT_FOLDER = "generated_reports"
