+++ airbyte-ci/connectors_ci/metadata_service/orchestrator/jobs/catalog.py
@@ -1,0 +1,11 @@
+
+from dagster import define_asset_job
+
+# ------ Jobs ------ #
+
+generate_catalog_markdown = define_asset_job(name="generate_catalog_markdown", selection="*")
+
+# ------ Sensors ------ #
+
+
+
