+++ airbyte-ci/connectors_ci/metadata_service/orchestrator/utils/html.py
@@ -1,0 +1,12 @@
+def html_body(title, content):
+    return f"""
+    <!DOCTYPE html>
+    <html>
+        <head>
+            <title>{title}</title>
+        </head>
+        <body>
+            {content}
+        </body>
+    </html>
+    """
